Split Renderer to Canvas
